made render general
