+++ src/components/layout.py
@@ -1,15 +1,14 @@
 from dash import Dash, html, dcc
 
 from ..tabs import (
-    prot_tab1,
-    rna_raw_tab,
+    box_tabs,
     rna_processed_tab,
     text_tab1,
     text_tab2,
-    phos_tab1,
 )
 
-from src.helpers import IDs
+from src.helpers import IDs, Params
+
 
 def tab_layout(name, subtabs_id, *args):
     return dcc.Tab(
@@ -21,7 +20,11 @@ def tab_layout(name, subtabs_id, *args):
 
 def sub_tab_layout(app, data, name, rendering, class_name="control-tab"):
     ids = IDs(name)
-    print(name)
+    params = (
+        Params(X="gene", COLOUR="ID", LOG=True, Y="abun")
+        if name == "Phosphoproteomics"
+        else Params(X="test")
+    )
     return dcc.Tab(
         label=name,
         value=name,
@@ -35,7 +38,7 @@ def sub_tab_layout(app, data, name, rendering, class_name="control-tab"):
                         html.Hr(),
                         html.Div(
                             className=class_name,
-                            children=[rendering.render(app, data, ids)],
+                            children=[rendering.render(app, data, ids, params)],
                         ),
                     ],
                 )
@@ -63,22 +66,24 @@ def create_layout(app: Dash, data) -> html.Div:
                         tab_layout(
                             "RNAseq (bulk)",
                             "subtabs_id1",
-                            sub_tab_layout(app, data, "CPM", rna_raw_tab),
+                            sub_tab_layout(app, data["rna_bulk"], "CPM", box_tabs),
                             sub_tab_layout(
-                                app, data, "Volcano", rna_processed_tab
+                                app, data["rna_bulk"], "Volcano", rna_processed_tab
                             ),
                         ),
                         tab_layout("RNA_single_cell", "subtabs_id2"),
-                        # tab_layout(
-                        #     "Protein",
-                        #     "subtabs_id3",
-                        #     sub_tab_layout(app, data, "Proteins", prot_tab1),
-                        # ),
                         tab_layout(
                             "Phospho/proteomics",
                             "subtabs_id4",
-                            sub_tab_layout(app, data, "Proteins", prot_tab1),
-                            sub_tab_layout(app, data, "Phosphoproteomics", phos_tab1),
+                            sub_tab_layout(
+                                app, data["proteomics"], "Proteins", box_tabs
+                            ),
+                            sub_tab_layout(
+                                app,
+                                data["phosphoproteomics"],
+                                "Phosphoproteomics",
+                                box_tabs,
+                            ),
                         ),
                         tab_layout("Function", "subtabs_id5"),
                     ],
