+++ src/helpers.py
@@ -1,15 +1,8 @@
 import numpy as np
-from time import time
-import pandas as pd
 from dash import dcc, html
 import plotly.express as px
 from src.read_files import Data
-
-# from .components import ids
-from dash.dependencies import Input, Output
-from typing import Any
 from dataclasses import dataclass
-
 
 @dataclass
 class Params:
@@ -36,6 +29,10 @@ class IDs:
         return f"{self.name}_gene-dropdown"
 
     @property
+    def plot(self) -> str:
+        return f"{self.name}_plot"
+
+    @property
     def tests_drop(self) -> str:
         return f"{self.name}_tests-dropdown"
 
@@ -44,16 +41,17 @@ class IDs:
         return f"{self.name}_select-all"
 
     @property
-    def plot(self) -> str:
-        return f"{self.name}_plot"
-
-    @property
     def slider1(self) -> str:
         return f"{self.name}_slider1"
 
     @property
     def slider2(self) -> str:
         return f"{self.name}_slider2"
+
+
+# class TestIDs(IDs):
+# class VolcanoIDs(IDs):
+
 
 
 def make_list_of_dicts(values: list[str]) -> list[dict[str, str]]:
@@ -222,97 +220,3 @@ def draw_box_chart(data: Data, y_gene: str, params: type, plot_id: str) -> html.
             fig.update_layout(margin=dict(t=i * 33))
 
     return html.Div(dcc.Graph(figure=fig), id=plot_id)
-
-
-def gene_dropdown(app, ids: IDs, data_set: Data):
-    @app.callback(Output(ids.gene_drop, "options"), Input(ids.data_drop, "value"))
-    def set_gene_options(experiment: str) -> list[dict[str, str]]:
-        """Populates the gene selection dropdown with options from teh given dataset"""
-        return make_list_of_dicts(list(data_set[experiment].df.columns))
-
-
-def gene_dropdown_default(app, ids: IDs):
-    @app.callback(Output(ids.gene_drop, "value"), Input(ids.gene_drop, "options"))
-    def select_gene_value(gene_options: list[dict[str, str]]) -> str:
-        """Select first gene as default value"""
-        return gene_options[0]["value"]
-
-
-def box(app, ids: IDs, data_set: Data, params: type):
-    @app.callback(
-        Output(ids.plot, "children"),
-        Input(ids.data_drop, "value"),
-        Input(ids.gene_drop, "value"),
-        Input(ids.tests_drop, "value"),
-    )
-    def update_box_chart(experiment: str, gene: str, tests: list[str]) -> html.Div:
-        """Re draws a box and wisker of the CPM data for each set of replicates for eact
-        test and overlays the respective FDR value"""
-        selected_data = data_set[experiment]
-        filtered: Data = selected_data.filter(gene, tests)
-        y_param = params.Y if params.Y else gene
-        return draw_box_chart(filtered, y_param, params, ids.plot)
-
-
-def test_dropdown(app, ids: IDs, data_set: Data):
-    @app.callback(
-        Output(ids.tests_drop, "options"),
-        Input(ids.data_drop, "value"),
-    )
-    def set_comparison_options(experiment: str) -> list[dict[str, str]]:
-        """Populates the test selection dropdown with options from teh given dataset"""
-        return make_list_of_dicts(list(data_set[experiment].test_names))
-
-
-def test_dropdown_select_all(app, ids: IDs):
-    @app.callback(
-        Output(ids.tests_drop, "value"),
-        Input(ids.tests_drop, "options"),
-        Input(ids.select_all, "n_clicks"),
-    )
-    def select_comparison_values(
-        available_comparisons: list[dict[str, str]], _: int
-    ) -> list[dict[str, str]]:
-        """Default to all available comparisons"""
-        return [comp["value"] for comp in available_comparisons]
-
-
-def get_defaults(data: Data, key: str) -> tuple[str, Data]:
-    datasets = list(data[key].keys())
-    first_dataset = data[key][datasets[0]]
-    first_gene = first_dataset.df.columns[0]
-    dataset = first_dataset.filter(first_gene, list(first_dataset.test_names))
-
-    return first_gene, dataset, datasets
-
-
-def dropdowns(data: Data, key: str, params: Params, ids: IDs) -> list[Any]:
-    first_gene, dataset, datasets = get_defaults(data, key)
-
-    children = [
-        html.H6("Dataset"),
-        dcc.Dropdown(
-            id=ids.data_drop,
-            options=datasets,
-            value=datasets[0],
-            multi=False,
-        ),
-        html.H6("Gene"),
-        dcc.Dropdown(
-            id=ids.gene_drop,
-        ),
-        html.H6("test"),
-        dcc.Dropdown(
-            id=ids.tests_drop,
-            multi=True,
-        ),
-        html.Button(
-            className="dropdown-button",
-            children=["Select All"],
-            id=ids.select_all,
-            n_clicks=0,
-        ),
-        html.Div(draw_box_chart(dataset, first_gene, params, ids.plot)),
-    ]
-
-    return children
