+++ src/tabs/text_tab1.py
@@ -3,7 +3,7 @@ from dash import Dash, html
 from src.read_files import RNASeqData
 
 
-def render(app: Dash, data: dict[str, RNASeqData], ids) -> html.Div:
+def render(app: Dash, data: dict[str, RNASeqData], ids, params) -> html.Div:
     return html.Div(
         children=[
             html.P("Wecome to the lab Hudson ploting dashbouard! "),
